pyproject.toml
[project]
name = "actors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
    "pyyaml>=6.0.2",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "ruff>=0.12.3",
    "pytest>=8.4.1",
    "bcrypt>=4.3.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.10.1",
    "pytest-asyncio>=1.1.0",
    "geopy>=2.4.1",
    "pytz>=2024.1",
    "requests>=2.32.0",
    "yfinance>=0.2.38",
]



